refactor(family model): Moved MonsterFamily models to separate file
- MonsterFamily related models in src/app/models/dqm1/models/monster.py moved to separate file within the same directory
